+++ src/core/datatypes/complex/container/Vector/VectorCharData/VectorCharacters.cpp
@@ -23,6 +23,7 @@
 #include "RbString.h"
 #include "VectorCharacters.h"
 
+#include <algorithm>
 
 
 /** Constructor with element type, used to properly construct vectors */
@@ -101,7 +102,7 @@ void VectorCharacters::unique(void) {
     sort();
     std::vector<RbLanguageObject*> uniqueVector;
     uniqueVector.push_back (elements[0]);
-    for (int i = 1 ; i< elements.size() ; i++)
+    for (size_t i = 1 ; i< elements.size() ; i++)
     {
         if (*(static_cast<Character*>(elements[i])) != *(static_cast<Character*>(elements[i-1])))
             uniqueVector.push_back(elements[i]);
