+++ src/core/datatypes/complex/container/Vector/VectorString.cpp
@@ -23,6 +23,7 @@
 #include "VectorInteger.h"
 #include "VectorString.h"
 
+#include <algorithm>
 #include <sstream>
 
 
@@ -202,7 +203,7 @@ void VectorString::unique(void) {
     sort();
     std::vector<RbLanguageObject*> uniqueVector;
     uniqueVector.push_back (elements[0]);
-    for (int i = 1 ; i< elements.size() ; i++)
+    for (size_t i = 1 ; i< elements.size() ; i++)
     {
         if (*(static_cast<RbString*>(elements[i])) != *(static_cast<RbString*>(elements[i-1])))
             uniqueVector.push_back(elements[i]);
